requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gitpython==3.1.40
langchain==0.3.25
langchain-community==0.3.24
langchain-groq>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma==0.2.4
chromadb>=1.0.9
sentence-transformers>=3.0.0
